Point float -> double

+++ app/src/main/java/int09h/facebook/com/sbtexhere3/models/Geo.kt
@@ -6,7 +6,7 @@ import org.json.JSONObject
  * Created by int09h on 23/02/2018.
  */
 
-data class Point (val latitude: Float, val longitude: Float)
+data class Point (val latitude: Double, val longitude: Double)
 
 data class Triangle (val left: Point, val right: Point, val cb: Point) {
 
@@ -33,9 +33,9 @@ data class Triangle (val left: Point, val right: Point, val cb: Point) {
 
     companion object {
         fun createFromPointAndRadius(point: Point, radius: Float): Triangle {
-            val left = Point(0.0f, 0.0f)
-            val right = Point(0.0f, 0.0f)
-            val cornerBack = Point(0.0f, 0.0f)
+            val left = Point(0.0, 0.0)
+            val right = Point(0.0, 0.0)
+            val cornerBack = Point(0.0, 0.0)
             return Triangle(left, right, cornerBack)
         }
     }
